Some improvements

+++ includes/class-import-eventbrite-events-list-table.php
@@ -537,10 +537,13 @@ class Import_Eventbrite_Events_History_List_Table extends WP_List_Table {
 			'iee_action' => 'iee_all_history_delete',
 		);
 		
-		$wp_delete_noonce_url = esc_url( wp_nonce_url( add_query_arg( $iee_url_all_delete_args ), 'iee_delete_all_history_nonce' ) );
-		$actions = '<a class="button apply" href="'.$wp_delete_noonce_url.'" onclick="return confirm(\'Warning!! Are you sure to Delete this import history? Import history will be permanatly deleted.\')">'.esc_html__( 'Clear Import History', 'import-eventbrite-events' ).'</a>';
-		echo $actions;
-		
+		$delete_ids  = get_posts( array( 'numberposts' => 1,'fields' => 'ids', 'post_type'   => 'iee_import_history' ) );
+		if( !empty( $delete_ids ) ){
+			$wp_delete_noonce_url = esc_url( wp_nonce_url( add_query_arg( $iee_url_all_delete_args, admin_url( 'admin.php' ) ), 'iee_delete_all_history_nonce' ) );
+			$actions = '<a class="button apply" href="'.$wp_delete_noonce_url.'" onclick="return confirm(\'Warning!! Are you sure to Delete this import history? Import history will be permanatly deleted.\')">'.esc_html__( 'Clear Import History', 'import-eventbrite-events' ).'</a>';
+			echo $actions;
+		}
+
 		$action = array(
 			'all_delete' => $actions,
 		);
